chore: factor group submission view into component
- #120
